Add some comments. Update version. Rename variables (#20)
Reset the version to 1.0.0 for first release.

Add comments to the InitSpeechConnector() method

Reorder settings in InitSpeechConnector(), so most used are at the top

Update two variable names (endpoint -==> botsceret) to match new terminology

Remove the option to embed a friendly name in the bot secret, to help memorize which bot this secret belongs to. I'll open an issue to find a better way to address this and document the solution.