Add filter for wakeword table file (#44)
Filter for .table files, to make sure users know that we can't directly consume the .zip that was downloaded from the Microsoft wake word web portal -- you first need to unzip it.
The unzipped file is always named kws.table (unfortunately). The file name does not reflect the selected wake word, so need to fix the message string to reflect that.